requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
discord.py==2.6.2
frozenlist==1.7.0
google-auth==2.40.3
google-genai==1.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.6.4
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
